testing/web-platform/mozilla/tests/binast/domain-restrict-excluded.https.html

<!DOCTYPE html>
<title>Check we can't load BinAST if the host is excluded in the list</title>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    setup({allow_uncaught_exception: true});

    var hadScriptLoadError = false;
    function setLoadError() {
      // Load error happens if the server side throws an exception,
      // for 'expect_accept' check on server side.
      hadScriptLoadError = true;
    }

    var hadSyntaxError = false;
    var hadOtherError = false;
    function checkSyntaxError(event) {
      // Server returns BinAST and the browser parses it as plain JS.
      if (event.message.startsWith("SyntaxError:")) {
        hadSyntaxError = true;
      } else {
        hadOtherError = true;
      }
    }

    window.addEventListener("error", checkSyntaxError);

    const test_load = async_test("Check we can't load BinAST if the host is excluded in the list");
    window.addEventListener("load", test_load.step_func_done(ev => {
      assert_equals(hadScriptLoadError, false, "Didn't expect a load error event");
      assert_equals(hadSyntaxError, true, "Expect a syntax error event for receiving binast");
      assert_equals(hadOtherError, false, "Didn't expect other error event");
      assert_equals(typeof binASTLoaded, "undefined", "Expected not to load either version");
    }));

</script>
<script src="./serve.py?name=small&amp;expect_accept=false&amp;force_binast=true" onerror="setLoadError()"></script>
